refactor: test, eat, signUp path wip
