+++ FrontEnd/src/pages/auth/form/Employee.jsx
@@ -1,10 +1,4 @@
 import React from 'react';
-import { CompanyDropdown, DropdownItem, InputBox, Label } from './sharedStyles';
-import {
-  createHandleChange,
-  createHandleCompanyNameKeyUp,
-  createHandleCompanySelect,
-} from './Handler';
 import styled from 'styled-components';
 
 const EmployeeStep = ({
@@ -15,13 +9,46 @@ const EmployeeStep = ({
   memberService,
   companySearchResults,
 }) => {
-  const handleChange = createHandleChange(() => {}, setFormData2);
-  const handleCompanyNameKeyUp = createHandleCompanyNameKeyUp(
-    companyNameTimeout,
-    setCompanySearchResults,
-    memberService
-  );
-  const handleCompanySelect = createHandleCompanySelect(setFormData2, setCompanySearchResults);
+  const handleChange = (e, step) => {
+    const { name, value } = e.target;
+    if (step === 1) setFormData1((prev) => ({ ...prev, [name]: value }));
+    else if (step === 2) setFormData2((prev) => ({ ...prev, [name]: value }));
+  };
+
+  const handleCompanyNameKeyUp = (e) => {
+    const value = e.target.value;
+    if (companyNameTimeout.current) clearTimeout(companyNameTimeout.current);
+    companyNameTimeout.current = setTimeout(async () => {
+      if (!value) {
+        setCompanySearchResults([]);
+        return;
+      }
+      try {
+        const data = await memberService.searchCompany(value);
+        setCompanySearchResults(Array.isArray(data) ? data : []);
+      } catch (err) {
+        setCompanySearchResults([]);
+        console.error('회사명 검색 에러:', err);
+      }
+    }, 500);
+  };
+
+  const handleCompanySelect = (company) => {
+    setFormData2((prev) => ({
+      ...prev,
+      worcation_no: company.company_no,
+      company_address: company.company_address,
+    }));
+    setCompanySearchResults([]);
+  };
+
+  const handleDepartmentSelect = () => {
+    // 부서 선택 로직
+  };
+
+  const handlePositionSelect = () => {
+    // 직급 선택 로직
+  };
 
   return (
     <Body className="InputWrap">
@@ -129,8 +156,76 @@ const Left = styled.div`
   flex-direction: column;
   align-items: flex-start;
 `;
+
 const Right = styled.div`
   display: flex;
   flex-direction: column;
   align-items: flex-start;
+`;
+
+const InputBox = styled.input`
+  ${({ variant }) => {
+    switch (variant) {
+      case 'yellow':
+        return `
+          background: #ffffff;
+          border: 3px solid #ffeb8c;
+          border-radius: 10px;
+          color: black;
+        `;
+      case 'gray':
+        return `
+          background: #f3f3f3;
+          border: 3px solid #d1d5db;
+          border-radius: 10px;
+          color: black;
+        `;
+      case 'orange':
+        return `
+          background: #ffffff;
+          border: 3px solid #f59e0b;
+          border-radius: 10px;
+          color: black;
+        `;
+      default:
+        return '';
+    }
+  }};
+  width: 400px;
+  padding: ${({ theme }) => theme.spacing.s3};
+  box-shadow: 4px 4px 4px 0 rgba(0, 0, 0, 0.25);
+  margin: ${({ theme }) => theme.spacing.s0};
+`;
+
+const Label = styled.label`
+  display: block;
+  text-align: left;
+  margin-bottom: ${({ theme }) => theme.spacing.s2};
+  margin-left: ${({ theme }) => theme.spacing.s1};
+`;
+
+const CompanyDropdown = styled.ul`
+  position: absolute;
+  top: 100%;
+  left: 0;
+  width: 100%;
+  background: #fffbe6;
+  border: 1px solid #ffeb8c;
+  border-radius: 0 0 10px 10px;
+  box-shadow: 0 2px 8px rgba(0, 0, 0, 0.08);
+  z-index: 10;
+  max-height: 220px;
+  overflow-y: auto;
+  margin: 0;
+  padding: 0;
+  list-style: none;
+`;
+
+const DropdownItem = styled.li`
+  padding: 10px 16px;
+  height: 44px;
+  cursor: pointer;
+  &:hover {
+    background: #fff3b0;
+  }
 `;
