+++ FrontEnd/src/pages/auth/form/Master.jsx
@@ -1,19 +1,4 @@
 import React, { useState } from 'react';
-import {
-  CompanyDropdown,
-  DropdownItem,
-  InputBox,
-  Label,
-  Btn,
-  AddressSearchButton,
-  StyledDatePickerWrapper,
-} from './sharedStyles';
-import {
-  createHandleChange,
-  createHandleCompanyNameKeyUp,
-  createHandleCompanySelect,
-  createHandleAddressSearch,
-} from './Handler';
 import styled from 'styled-components';
 import Popup from './Popup';
 import CustomDatePicker from '../../../components/common/DatePicker';
@@ -32,14 +17,51 @@ const MasterStep = ({
   const [showInput, setShowInput] = useState(false);
   const [departments, setDepartments] = useState([]);
 
-  const handleChange = createHandleChange(() => {}, setFormData2);
-  const handleCompanyNameKeyUp = createHandleCompanyNameKeyUp(
-    companyNameTimeout,
-    setCompanySearchResults,
-    memberService
-  );
-  const handleCompanySelect = createHandleCompanySelect(setFormData2, setCompanySearchResults);
-  const handleAddressSearch = createHandleAddressSearch(isPostcodeReady, setFormData2);
+  const handleChange = (e, step) => {
+    const { name, value } = e.target;
+    if (step === 1) setFormData1((prev) => ({ ...prev, [name]: value }));
+    else if (step === 2) setFormData2((prev) => ({ ...prev, [name]: value }));
+  };
+
+  const handleCompanyNameKeyUp = (e) => {
+    const value = e.target.value;
+    if (companyNameTimeout.current) clearTimeout(companyNameTimeout.current);
+    companyNameTimeout.current = setTimeout(async () => {
+      if (!value) {
+        setCompanySearchResults([]);
+        return;
+      }
+      try {
+        const data = await memberService.searchCompany(value);
+        setCompanySearchResults(Array.isArray(data) ? data : []);
+      } catch (err) {
+        setCompanySearchResults([]);
+        console.error('회사명 검색 에러:', err);
+      }
+    }, 500);
+  };
+
+  const handleCompanySelect = (company) => {
+    setFormData2((prev) => ({
+      ...prev,
+      worcation_no: company.company_no,
+      company_address: company.company_address,
+    }));
+    setCompanySearchResults([]);
+  };
+
+  const handleAddressSearch = () => {
+    if (!isPostcodeReady) {
+      alert('주소 검색 스크립트가 아직 준비되지 않았습니다.');
+      return;
+    }
+    new window.daum.Postcode({
+      oncomplete: function (data) {
+        const addr = data.address;
+        setFormData2((prev) => ({ ...prev, address: addr }));
+      },
+    }).open();
+  };
 
   const handleDeptBtnClick = () => {
     if (showInput) {
@@ -191,8 +213,126 @@ const Left = styled.div`
   flex-direction: column;
   align-items: flex-start;
 `;
+
 const Right = styled.div`
   display: flex;
   flex-direction: column;
   align-items: flex-start;
+`;
+
+const InputBox = styled.input`
+  ${({ variant }) => {
+    switch (variant) {
+      case 'yellow':
+        return `
+          background: #ffffff;
+          border: 3px solid #ffeb8c;
+          border-radius: 10px;
+          color: black;
+        `;
+      case 'gray':
+        return `
+          background: #f3f3f3;
+          border: 3px solid #d1d5db;
+          border-radius: 10px;
+          color: black;
+        `;
+      case 'orange':
+        return `
+          background: #ffffff;
+          border: 3px solid #f59e0b;
+          border-radius: 10px;
+          color: black;
+        `;
+      default:
+        return '';
+    }
+  }};
+  width: 400px;
+  padding: ${({ theme }) => theme.spacing.s3};
+  box-shadow: 4px 4px 4px 0 rgba(0, 0, 0, 0.25);
+  margin: ${({ theme }) => theme.spacing.s0};
+`;
+
+const Btn = styled.button`
+  width: 100px;
+  height: 52.41px;
+  background-color: ${({ theme }) => theme.colors.primary};
+  color: ${({ theme }) => theme.colors.black};
+  border: none;
+`;
+
+const Label = styled.label`
+  display: block;
+  text-align: left;
+  margin-bottom: ${({ theme }) => theme.spacing.s2};
+  margin-left: ${({ theme }) => theme.spacing.s1};
+`;
+
+const StyledDatePickerWrapper = styled.div`
+  .react-datepicker__input-container input {
+    box-sizing: border-box;
+    max-width: 100%;
+    width: 400px;
+    height: 52.41px;
+    background: #ffffff;
+    border: 3px solid #ffeb8c;
+    border-radius: 10px;
+    color: black;
+    padding: 0 12px;
+    font-size: ${({ theme }) => theme.fontSizes.base};
+    outline: none;
+    margin: 0;
+    box-shadow: 4px 4px 4px 0 rgba(0, 0, 0, 0.25);
+  }
+  .react-datepicker__input-container input::placeholder {
+    color: #adadad;
+    opacity: 1;
+  }
+  .react-datepicker__input-container input:focus {
+    border-color: #ffeb8c;
+    background: #fff;
+  }
+`;
+
+const CompanyDropdown = styled.ul`
+  position: absolute;
+  top: 100%;
+  left: 0;
+  width: 100%;
+  background: #fffbe6;
+  border: 1px solid #ffeb8c;
+  border-radius: 0 0 10px 10px;
+  box-shadow: 0 2px 8px rgba(0, 0, 0, 0.08);
+  z-index: 10;
+  max-height: 220px;
+  overflow-y: auto;
+  margin: 0;
+  padding: 0;
+  list-style: none;
+`;
+
+const DropdownItem = styled.li`
+  padding: 10px 16px;
+  height: 44px;
+  cursor: pointer;
+  &:hover {
+    background: #fff3b0;
+  }
+`;
+
+const AddressSearchButton = styled.button`
+  height: 52.41px;
+  min-width: 80px;
+  padding: 0 12px;
+  background: #feffe0;
+  border: 3px solid #dda900;
+  border-radius: ${({ theme }) => theme.borderRadius.lg};
+  font-family: ${({ theme }) => theme.fontFamily.primary};
+  font-size: ${({ theme }) => theme.fontSizes.base};
+  color: ${({ theme }) => theme.colors.black};
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  cursor: pointer;
 `;
