+++ FrontEnd/src/pages/eat/Eat.jsx
@@ -1,7 +1,7 @@
 import React, { useState } from 'react';
-import NonCompleteTest from './components/NonCompleteTest';
+import NonCompleteTest from './form/NonCompleteTest';
 import userStore from '../../store/userStore';
-import CompleteTest from './components/CompleteTest';
+import CompleteTest from './form/CompleteTest';
 
 const Eat = () => {
   const { body, stress, burnout, tendency } = userStore();
