+++ FrontEnd/src/pages/worcation/Register.jsx
@@ -1,14 +1,14 @@
 import React, { useState } from 'react';
 import styled from 'styled-components';
-import Menu from './element/Menu';
-import ApplicationForm from './element/form/ApplicationForm.jsx';
-import InfoForm from './element/form/InfoForm.jsx';
-import DescriptionForm from './element/form/DescriptionForm.jsx';
-import PhotoForm from './element/form/PhotoForm.jsx';
-import AmenitiesForm from './element/form/AmenitiesForm.jsx';
-import LocationForm from './element/form/LocationForm.jsx';
-import PolicyForm from './element/form/PolicyForm.jsx';
-import FeatureForm from './element/form/FeatureForm.jsx';
+import Menu from './Menu.jsx';
+import ApplicationForm from './form/ApplicationForm.jsx';
+import InfoForm from './form/InfoForm.jsx';
+import DescriptionForm from './form/DescriptionForm.jsx';
+import PhotoForm from './form/PhotoForm.jsx';
+import AmenitiesForm from './form/AmenitiesForm.jsx';
+import LocationForm from './form/LocationForm.jsx';
+import PolicyForm from './form/PolicyForm.jsx';
+import FeatureForm from './form/FeatureForm.jsx';
 import { BtnWhiteYellowBorder } from '../../styles/Button.styles';
 import Swal from 'sweetalert2';
 import SwalStyles from '../../styles/SwalStyles';
