+++ FrontEnd/src/pages/worcation/form/PhotoForm.jsx
@@ -1,9 +1,9 @@
 import React, { useRef, useState } from 'react';
 import styled from 'styled-components';
-import { InputLightGray } from '../../../../styles/Input.styles.js';
-import AddButton from '../../../../components/common/AddButton.jsx';
-import ImageUploader from '../../../../components/common/ImageUploader.jsx';
-import { ButtonBorder } from '../../../../styles/Button.styles.js';
+import { InputLightGray } from '../../../styles/Input.styles.js';
+import AddButton from '../../../components/common/AddButton.jsx';
+import ImageUploader from '../../../components/common/ImageUploader.jsx';
+import { ButtonBorder } from '../../../styles/Button.styles.js';
 
 const Form = () => {
   const [officeImages, setOfficeImages] = useState([]);
